refactoring input parameter and controller logic
