+++ server/controllers/scriptController.js
@@ -1,52 +1,41 @@
-const puppeteer = require('puppeteer');
-const pathToExtension = require('path').join(__dirname, '../../extension');
-const scripts = require('../userscripts.js');
+const fs = require('fs');
 
-// RUN THIS TO RECORD USER ACTIONS FROM BROWSER AND CREATE PUPPET SCRIPT
-const createScript = async (name, inputURL) => {
-  if (scripts[name]) {
-    console.log('That script name already exists. Please try a new one.');
-    process.exit(0);
-  } else {
-    const browser = await puppeteer.launch({
-      headless: false,
-      args: [
-        `--disable-extensions-except=${pathToExtension}`,
-        `--load-extension=${pathToExtension}`,
-      ],
-    });
-    const page = (await browser.pages())[0];
-    await page.goto(inputURL);
-    await page.content();
+// the name of the script to be saved or executed
+const scriptName = process.argv[3];
+
+// take the URL to open in Puppeteer from the input script parameter
+let inputURL;
+if (process.argv[4]) {
+  inputURL = process.argv[4];
+}
+
+exports.storeScript = (req, res, next) => {
+  let input = req.body.code;
+  // sanitize the puppeteer script sent from the chrome extension
+  // remove the first two lines containing the URL and viewport information
+  let newString = input.slice(input.indexOf(`)`) + 1);
+  // replace blank lines with semi-colons
+  newString = newString
+    .slice(newString.indexOf(`)`) + 3)
+    .replace(/\)\n/g, ');');
+  // build an object containing the url and sanitized puppeteer script
+  let newScript = `exports.${scriptName} = {
+    url: '${inputURL}',
+    func: async (page) => {${newString}  }
   }
-};
 
-// RUN THIS AFTER RECORDING PUPPET SCRIPT
-// launch puppeteer headless and open the page provided by the user
-const runScript = async script => {
-  console.log('RUN SCRIPT');
-  let browser = await puppeteer.launch();
-  let context = await browser.createIncognitoBrowserContext();
-  let page = await context.newPage();
-  await page.goto(scripts[script].url);
-  await page.content();
-
-  let count = 0;
-
-  try {
-    while (count < 7) {
-      // Do something a couple of times. (insert recorded puppeteer scripts here)
-      scripts[script].func(page);
-      count++;
-    }
-  } catch (err) {
-    console.log('ERROR: ', err);
-  }
-  browser.close();
-  process.exit(0);
-};
-
-module.exports = {
-  createScript,
-  runScript,
-};
+  `;
+  // add the object locally to the userscripts file
+  fs.appendFile(
+    path.join(__dirname, './userscripts.js'),
+    newScript,
+    'utf-8',
+    function(err) {
+      if (err) next(err);
+      console.log(
+        `Saved successfully! You can run this test by entering "npm start -- run ${scriptName}"`,
+      );
+      next();
+    },
+  );
+}
