+++ server/userscripts.js
@@ -1,14 +1,5 @@
-exports.startStop = {
-  url: 'http://localhost:8080',
-  func: async page => {
-    await page.waitForSelector('body > #leak-button');
-    await page.click('body > #leak-button');
-    await page.waitForSelector('body > #stop-button');
-    await page.click('body > #stop-button');
-  },
-};
-exports.start = { 
-    url: 'http://localhost:8080',
+exports.newScript = { 
+    url: 'http://localhost:3000',
     func: async (page) => { 
   await page.waitForSelector('body > #leak-button');  await page.click('body > #leak-button');  
   }
